requirements.txt
requests==2.28.2
pandas==1.5.3
matplotlib==3.7.1
kafka-python==2.0.2
flink-python==1.16.0
numpy
requests
json